S3 method for summarizing prepare results

This function returns long summary of the athletemonitoring object

# S3 method for athletemonitoring
summary(object, ...)

Arguments

object
Object of class athletemonitoring

...
Extra arguments. Not used

Value

Tibble with athlete - variable summaries

Examples

# Load monitoring data set
data("monitoring")

# Filter out only 'Training Load'
monitoring <- monitoring[monitoring$Variable == "Training Load", ]

# Convert column to date format (or use numeric)
monitoring$Date <- as.Date(monitoring$Date, "%Y-%m-%d")

# Run the athlete monitoring data preparation
prepared_data <- prepare(
  data = monitoring,
  athlete = "Full Name",
  date = "Date",
  variable = "Variable",
  value = "Value",
  acute = 7,
  chronic = 42,

  # How should be missing entry treated?
  # What do we assume? Zero load? Let's keep NA
  NA_session = NA,

  # How should missing days (i.e. no entries) be treated?
  # Here we assume no training, hence zero
  NA_day = 0,

  # How should be multiple day entries summarised?
  # With "load", it is a "sum", witho other metrics that
  # do not aggregate, it can me "mean"
  day_aggregate = function(x) {
    sum(x, na.rm = TRUE)
  },

  # Rolling estimators for Acute and Chronic windows
  rolling_estimators = function(x) {
    c(
      "mean" = mean(x, na.rm = TRUE),
      "sd" = sd(x, na.rm = TRUE),
      "cv" = sd(x, na.rm = TRUE) / mean(x, na.rm = TRUE)
    )
  },

  # Additional estimator post-rolling
  posthoc_estimators = function(data) {
    data$ACD <- data$acute.mean - data$chronic.mean
    data$ACR <- data$acute.mean / data$chronic.mean
    data$ES <- data$ACD / data$chronic.sd

    # Make sure to return the data
    return(data)
  },

  # Group summary estimators
  group_summary_estimators = function(x) {
    c(
      "median" = median(x, na.rm = TRUE),
      "lower" = quantile(x, 0.25, na.rm = TRUE)[[1]],
      "upper" = quantile(x, 0.75, na.rm = TRUE)[[1]]
    )
  }
)

#> Preparing data...

#> Rolling...

#> Group summaries...

#> Missing data summaries...

#> Done!

# Get summary
prepared_data

#> Athlete monitoring numeric data with the following characteristics:
#> 
#> 10 athletes:
#> Alan McDonald, Ann Whitaker, Eve Black, Frank West, John Doe, Michael Peterson, Mike Smith, Peter Jackson, Stuart Rogan, Susan Kane 
#> 
#> 363 days:
#> From 18263 to 18625 
#> 
#> 0 missing entries
#> 510 missing days
#> 
#> 1 variables:
#> Training Load 
#> 
#> 10 estimators:
#> variable.value, acute.mean, acute.sd, acute.cv, chronic.mean, chronic.sd, chronic.cv, ACD, ACR, ES

summary(prepared_data)

#> # A tibble: 10 x 14
#>    athlete   variable `Day entries` `Missing entrie… `Missing days` `Start date`
#>    <chr>     <chr>            <int>            <dbl>          <dbl> <date>      
#>  1 Alan McD… Trainin…           363                0             51 2020-01-02  
#>  2 Ann Whit… Trainin…           363                0             51 2020-01-02  
#>  3 Eve Black Trainin…           363                0             51 2020-01-02  
#>  4 Frank We… Trainin…           363                0             51 2020-01-02  
#>  5 John Doe  Trainin…           363                0             51 2020-01-02  
#>  6 Michael … Trainin…           363                0             51 2020-01-02  
#>  7 Mike Smi… Trainin…           363                0             51 2020-01-02  
#>  8 Peter Ja… Trainin…           363                0             51 2020-01-02  
#>  9 Stuart R… Trainin…           363                0             51 2020-01-02  
#> 10 Susan Ka… Trainin…           363                0             51 2020-01-02  
#> # … with 8 more variables: Stop date <date>, Mean <dbl>, SD <dbl>, Min <dbl>,
#> #   Max <dbl>, Median <dbl>, IQR <dbl>, MAD <dbl>